+++ app/components/mark_complete_component/view.html.erb
@@ -1,0 +1,4 @@
+<%= form_with model: @mark_complete_form, url: @path, builder: GOVUKDesignSystemFormBuilder::FormBuilder do |f| %>
+  <%= f.govuk_collection_radio_buttons :mark_complete, @mark_complete_options, :value, :name, legend: { size: 'm' } %>
+  <%= f.govuk_submit %>
+<% end %>
